# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "InsightExtract API",
        "description": "API for PDF-to-flashcard conversion with spaced repetition",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/documents/{document_id}/decks": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "Get Document Decks",
                "description": "Get all decks for a document.",
                "operationId": "get_document_decks_api_documents__document_id__decks_get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DeckResponse"
                                    },
                                    "title": "Response Get Document Decks Api Documents  Document Id  Decks Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/decks": {
            "get": {
                "tags": [
                    "decks"
                ],
                "summary": "List Decks",
                "description": "List all decks with their card counts.",
                "operationId": "list_decks_api_decks_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/DeckResponse"
                                    },
                                    "type": "array",
                                    "title": "Response List Decks Api Decks Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "decks"
                ],
                "summary": "Create Deck",
                "description": "Create a new deck with cards.",
                "operationId": "create_deck_api_decks_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeckCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeckResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/decks/{deck_id}": {
            "get": {
                "tags": [
                    "decks"
                ],
                "summary": "Get Deck",
                "description": "Get deck by ID.",
                "operationId": "get_deck_api_decks__deck_id__get",
                "parameters": [
                    {
                        "name": "deck_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Deck Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeckResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/decks/{deck_id}/cards": {
            "get": {
                "tags": [
                    "decks"
                ],
                "summary": "Get Deck Cards",
                "description": "Get all cards in a deck.",
                "operationId": "get_deck_cards_api_decks__deck_id__cards_get",
                "parameters": [
                    {
                        "name": "deck_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Deck Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CardResponse"
                                    },
                                    "title": "Response Get Deck Cards Api Decks  Deck Id  Cards Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/decks/{deck_id}/export": {
            "post": {
                "tags": [
                    "decks"
                ],
                "summary": "Export Deck",
                "description": "Export deck to CSV or Anki format.",
                "operationId": "export_deck_api_decks__deck_id__export_post",
                "parameters": [
                    {
                        "name": "deck_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Deck Id"
                        }
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "csv",
                            "title": "Format"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Export Deck Api Decks  Deck Id  Export Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/jobs/{job_id}": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Get Job",
                "description": "Get job status by ID.",
                "operationId": "get_job_api_jobs__job_id__get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Job Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/{job_id}/stream": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Stream Job Progress",
                "description": "Stream job progress via Server-Sent Events.",
                "operationId": "stream_job_progress_api_jobs__job_id__stream_get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Job Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CardCreateInput": {
                "properties": {
                    "front": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Front"
                    },
                    "back": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Back"
                    },
                    "type": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FlashcardType"
                            }
                        ],
                        "default": "qa"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "front",
                    "back"
                ],
                "title": "CardCreateInput",
                "description": "Card creation input for deck creation."
            },
            "CardResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "deck_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Deck Id"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FlashcardType"
                    },
                    "front": {
                        "type": "string",
                        "title": "Front"
                    },
                    "back": {
                        "type": "string",
                        "title": "Back"
                    },
                    "citations": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Citations"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "ease": {
                        "type": "number",
                        "title": "Ease"
                    },
                    "interval": {
                        "type": "integer",
                        "title": "Interval"
                    },
                    "repetition": {
                        "type": "integer",
                        "title": "Repetition"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Due Date"
                    },
                    "last_reviewed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Reviewed At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "deck_id",
                    "type",
                    "front",
                    "back",
                    "citations",
                    "tags",
                    "ease",
                    "interval",
                    "repetition",
                    "due_date"
                ],
                "title": "CardResponse",
                "description": "Card response schema."
            },
...
            "DeckCreate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 1,
                        "title": "Title"
                    },
                    "documentId": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Documentid"
                    },
                    "cards": {
                        "items": {
                            "$ref": "#/components/schemas/CardCreateInput"
                        },
                        "type": "array",
                        "title": "Cards",
                        "default": []
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "title"
                ],
                "title": "DeckCreate",
                "description": "Deck creation schema."
            },
            "DeckResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "document_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Document Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "card_count": {
                        "type": "integer",
                        "title": "Card Count"
                    },
                    "due_count": {
                        "type": "integer",
                        "title": "Due Count"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "card_count",
                    "due_count",
                    "tags",
                    "created_at"
                ],
                "title": "DeckResponse",
                "description": "Deck response schema."
            },
...
            "FlashcardType": {
                "type": "string",
                "enum": [
                    "qa",
                    "cloze",
                    "truefalse"
                ],
                "title": "FlashcardType",
                "description": "Flashcard type enum."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "JobResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "document_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Document Id"
                    },
                    "type": {
                        "$ref": "#/components/schemas/JobType"
                    },
                    "status": {
                        "$ref": "#/components/schemas/JobStatus"
                    },
                    "progress": {
                        "type": "integer",
                        "title": "Progress"
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "completed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "document_id",
                    "type",
                    "status",
                    "progress",
                    "created_at"
                ],
                "title": "JobResponse",
                "description": "Job response schema."
            },
            "JobStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "running",
                    "completed",
                    "failed"
                ],
                "title": "JobStatus",
                "description": "Job status enum."
            },
            "JobType": {
                "type": "string",
                "enum": [
                    "extract",
                    "embed",
                    "summarize",
                    "cards"
                ],
                "title": "JobType",
                "description": "Job type enum."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}